requirements.txt
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
